// web/src/api/models/merge-suggestions-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { MergeSuggestionItem } from './merge-suggestion-item';

/**
 * Response containing node merge suggestions
 * @export
 * @interface MergeSuggestionsResponse
 */
export interface MergeSuggestionsResponse {
    /**
     * List of merge suggestions ordered by confidence score (highest first)
     * @type {Array<MergeSuggestionItem>}
     * @memberof MergeSuggestionsResponse
     */
    'suggestions': Array<MergeSuggestionItem>;
    /**
     * Total number of nodes analyzed
     * @type {number}
     * @memberof MergeSuggestionsResponse
     */
    'total_analyzed_nodes': number;
    /**
     * Processing time in seconds
     * @type {number}
     * @memberof MergeSuggestionsResponse
     */
    'processing_time_seconds': number;
    /**
     * Whether suggestions were loaded from cache
     * @type {boolean}
     * @memberof MergeSuggestionsResponse
     */
    'from_cache': boolean;
    /**
     * Generation timestamp
     * @type {string}
     * @memberof MergeSuggestionsResponse
     */
    'generated_at': string;
    /**
     * Total number of suggestions
     * @type {number}
     * @memberof MergeSuggestionsResponse
     */
    'total_suggestions': number;
    /**
     * Number of pending suggestions
     * @type {number}
     * @memberof MergeSuggestionsResponse
     */
    'pending_count': number;
    /**
     * Number of accepted suggestions
     * @type {number}
     * @memberof MergeSuggestionsResponse
     */
    'accepted_count': number;
    /**
     * Number of rejected suggestions
     * @type {number}
     * @memberof MergeSuggestionsResponse
     */
    'rejected_count': number;
}